move titan into titan.client package
