+++ app/src/main/java/titan/client/Client.java
@@ -1,4 +1,4 @@
-package titan;
+package titan.client;
 
 /**
  * Titan client interface: https://github.com/nb-titan/titan
