+++ app/src/main/java/titan/client/ClientImpl.java
@@ -1,4 +1,4 @@
-package titan;
+package titan.client;
 
 /**
  * Titan client implementation: https://github.com/nb-titan/titan
